EMS.BLL/Services/DepartmentServices.cs
using EMS.BLL.IServices;
using EMS.DAL.IRepository;
using EMS.Infrastructure.Entities;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace EMS.BLL.Services
{
    public class DepartmentServices : IDepartmentServices
    {
        private readonly IDepartmentRepository _departmentRepository;
        public DepartmentServices(IDepartmentRepository departmentRepository)
        {
            _departmentRepository = departmentRepository;
        }
        public async Task<List<Department>> GetDepartments()
            => await _departmentRepository.GetDepartments();
    }
}
